refactor(program-ops): split roster and events tabs out of program detail page
page.tsx was 605 lines with 10+ form-state vars and 6+ async handlers for
three unrelated tab panels. Roster (volunteer/participant CRUD) and Events
(read-only table) had no cross-tab state dependencies, so they extract
cleanly into sibling components; General tab stays inline since its state
is too entangled to extract without a 15-prop interface.

Roster's saving/picker state now lives locally in ProgramRosterTab instead
of sharing the page's single `saving` flag with the General tab's save
button.
